requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cytoolz==0.12.3
dateparser==1.2.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
Flask==3.0.3
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
MarkupSafe==2.1.5
multidict==6.0.5
parsimonious==0.10.0
protobuf==5.26.1
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
toolz==0.12.1
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.1
web3==6.19.0
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
